# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PurrFect Spots API",
    "description": "\n        PurrFect Spots API helps you share and discover cat locations in a socially-connected way.\n\n        ### 🌐 API Specification\n        All endpoints follow the `/api/v1/` prefix convention for stable identification.\n\n        ### ✨ Key Features\n        * 📍 **Geo-Location Awareness**: High-precision cat spot sharing with privacy protection.\n        * 🤳 **AI Photo Intelligence**: Automated cat detection and visual attribute extraction.\n        * 🔒 **Enterprise-Grade Security**: PKCE-enabled OAuth 2.0 and Supabase-backed authentication.\n        * ⚡ **Performance Optimized**: Built-in rate limiting, GZip compression, and ETag support.\n        * 🔄 **Real-time Ready**: Designed for low-latency interactions and live updates.\n        ",
    "contact": {
      "name": "Purrfect Spots Team",
      "email": "support@purrfectspots.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/upload/quota": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "Get Upload Quota",
        "description": "Get current user upload quota status.",
        "operationId": "get_upload_quota_api_v1_upload_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadQuotaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/upload/cat": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload Cat Photo",
        "description": "Upload cat photo with location information.\n\nSecurity features:\n- Rate limited: 5 requests per minute\n- Magic bytes validation for file type\n- Input sanitization for all text fields\n- Security event logging\n\nThe image is automatically optimized (resized/compressed) before upload to S3.\n\nRaises:\n    HTTPException: 400 - If no cats are detected in image.\n    HTTPException: 429 - If daily upload limit is reached.\n    HTTPException: 500 - If image processing or upload fails.",
        "operationId": "upload_cat_photo_api_v1_upload_cat_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_cat_photo_api_v1_upload_cat_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_cat_photo_api_v1_upload_cat_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "lat": {
            "type": "string",
            "title": "Lat"
          },
          "lng": {
            "type": "string",
            "title": "Lng"
          },
          "location_name": {
            "type": "string",
            "title": "Location Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "cat_detection_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cat Detection Data"
          },
          "location_blurred": {
            "type": "string",
            "title": "Location Blurred",
            "default": "false"
          }
        },
        "type": "object",
        "required": [
          "file",
          "lat",
          "lng",
          "location_name"
        ],
        "title": "Body_upload_cat_photo_api_v1_upload_cat_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadQuotaResponse": {
        "properties": {
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "is_pro": {
            "type": "boolean",
            "title": "Is Pro"
          },
          "reset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reset Type"
          },
          "resets_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resets At"
          }
        },
        "type": "object",
        "required": [
          "used",
          "limit",
          "remaining",
          "is_pro"
        ],
        "title": "UploadQuotaResponse",
        "description": "Upload quota status response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}